/feat: criação do tratamento de exceção do endpoint deletePost

+++ Desafio2_MicroserviceB/src/main/java/com/compass/Desafio2_MicroserviceB/service/PostService.java
@@ -60,6 +60,11 @@ public class PostService {
     }
 
     public void deletePost(Long id) {
+        if (!postRepository.existsById(id)) {
+            throw new EntityNotFoundException("Post não encontrado");
+        }
+
+
         postRepository.deleteById(id);
     }
